edit errors and returns

+++ internal/domain/event/event.go
@@ -1,7 +1,7 @@
 package event
 
 type Event struct {
-	Id   uint	`json:"id" db:"id,omitempty"`
+	Id   int64	`json:"id" db:"id,omitempty"`
 	Name string `json:"name" db:"name"`
 	Location string `json:"location" db:"location"`
 }
